Deduplicate GridTransform/ValuesTransform strategy impls and error formatting
Macro-izes the near-identical GridTransform/ValuesTransform Strategy1D/2D/3D
impls (grid_transform_strategy_impl!/values_transform_strategy_impl!, ND
stays hand-written), hoists empty_cache to one shared definition, adds
ValuesTransform::cached_values_view to replace repeated
into_dimensionality().expect() call sites, and rewrites error.rs's
multi-failure formatters to use iterator chains instead of hand-rolled loops.
No behavior changes; addresses review findings from issue #56.
