+++ src/interpolator/three/strategies.rs
@@ -323,321 +323,12 @@ where
     }
 }
 
-impl<D, S> Strategy3D<D> for GridTransform<D::Elem, S>
-where
-    D: Data + RawDataClone + Clone,
-    D::Elem: Float + Debug,
-    S: for<'a> Strategy3D<ViewRepr<&'a D::Elem>> + Clone + Debug,
-{
-    /// Checks the axis count (must be 1 or 3) and that every raw grid coordinate is
-    /// in its axis's configured transform's domain.
-    fn validate(&self, data: &InterpData3DBase<D>) -> Result<(), ValidateError> {
-        self.transforms
-            .validate_len(3, "GridTransform", "transforms")?;
-        let transformed_grid: Vec<Array1<D::Elem>> = data
-            .grid
-            .iter()
-            .enumerate()
-            .map(|(dim, grid)| self.transform_axis(dim, grid.view()))
-            .collect::<Result<_, _>>()?;
-        let values = self.transformed_values_view(data.values.view());
-        let view = InterpData3DView {
-            grid: core::array::from_fn(|i| transformed_grid[i].view()),
-            values,
-        };
-        self.inner.validate(&view)
-    }
-
-    /// Transforms the grid into `grid_cache`, then initializes `inner` against a
-    /// transient view zipping `grid_cache` with `data.values`.
-    ///
-    /// A raw grid is always strictly increasing, but a decreasing transform (e.g.
-    /// `Reciprocal`) would otherwise leave that axis of `grid_cache` decreasing;
-    /// `Transform::is_increasing` flags that case so the transformed axis (and the
-    /// matching `values` axis) can be reversed back to ascending, matching every
-    /// downstream strategy's ascending-grid assumption.
-    fn init(&mut self, data: &InterpData3DBase<D>) -> Result<(), ValidateError> {
-        self.grid_cache = data
-            .grid
-            .iter()
-            .enumerate()
-            .map(|(dim, grid)| self.transform_axis(dim, grid.view()))
-            .collect::<Result<_, _>>()?;
-        let values = self.transformed_values_view(data.values.view());
-        let view = InterpData3DView {
-            grid: core::array::from_fn(|i| self.grid_cache[i].view()),
-            values,
-        };
-        self.inner.init(&view)
-    }
-
-    fn interpolate(
-        &self,
-        data: &InterpData3DBase<D>,
-        point: &[D::Elem; 3],
-    ) -> Result<D::Elem, InterpolateError> {
-        self.check_point_domain(point)?;
-        let transformed_point: [D::Elem; 3] =
-            core::array::from_fn(|i| self.transforms[i].forward(point[i]));
-        let values = self.transformed_values_view(data.values.view());
-        let view = InterpData3DView {
-            grid: core::array::from_fn(|i| self.grid_cache[i].view()),
-            values,
-        };
-        self.inner.interpolate(&view, &transformed_point)
-    }
-
-    /// Forward-transforms into `grid_cache`'s coordinate space, then delegates the
-    /// actual wrap to `inner.interpolate_wrapped` rather than wrapping here itself:
-    /// wrapping doesn't commute with a nonlinear transform, so it must happen in the
-    /// *final* (innermost) transformed space where the periodic strategy actually
-    /// lives, mirroring how `ValuesTransform::interpolate_wrapped` defers to `inner`
-    /// for the same reason (composing two `GridTransform`s and wrapping at the outer
-    /// layer's space, then forward-transforming again, uses the wrong period). For a
-    /// non-transform `inner`, `Strategy3D::interpolate_wrapped`'s default wraps
-    /// directly against `grid_cache`, reproducing this layer's own wrap exactly.
-    fn interpolate_wrapped(
-        &self,
-        data: &InterpData3DBase<D>,
-        point: &[D::Elem; 3],
-    ) -> Result<D::Elem, InterpolateError>
-    where
-        D::Elem: Num + Euclid + Copy,
-    {
-        self.check_point_domain(point)?;
-        let transformed_point: [D::Elem; 3] =
-            core::array::from_fn(|i| self.transforms[i].forward(point[i]));
-        let values = self.transformed_values_view(data.values.view());
-        let view = InterpData3DView {
-            grid: core::array::from_fn(|i| self.grid_cache[i].view()),
-            values,
-        };
-        self.inner.interpolate_wrapped(&view, &transformed_point)
-    }
-
-    /// Skips the domain check `interpolate` does, and calls `inner.interpolate_fast`
-    /// rather than `inner.interpolate`, so "fast" propagates through nested
-    /// `GridTransform`/`ValuesTransform` layers instead of stopping at the outermost
-    /// one.
-    ///
-    /// An out-of-domain point is the caller's problem here, same as any other
-    /// unchecked `_fast` method: `forward`-ing it produces `NaN` (`Log`/`Sqrt`) or
-    /// `+-inf` (`Reciprocal`, only at `x == 0`). `NaN` poisons the grid search's
-    /// comparisons and panics; `+-inf` compares normally, so it's treated like an
-    /// ordinary out-of-bounds extrapolation query and silently produces `NaN` output
-    /// instead. Expected, not a bug: check [`Transform::in_domain`] yourself first if
-    /// you need a guarantee either way.
-    fn interpolate_fast(&self, data: &InterpData3DBase<D>, point: &[D::Elem; 3]) -> D::Elem {
-        let transformed_point: [D::Elem; 3] =
-            core::array::from_fn(|i| self.transforms[i].forward(point[i]));
-        let values = self.transformed_values_view(data.values.view());
-        let view = InterpData3DView {
-            grid: core::array::from_fn(|i| self.grid_cache[i].view()),
-            values,
-        };
-        self.inner.interpolate_fast(&view, &transformed_point)
-    }
-
-    /// Domain-checks every point in the batch before transforming, aggregating
-    /// every violation across the *whole batch* into one
-    /// [`InterpolateError::GridTransformDomain`] instead of erroring on the first
-    /// one, mirroring how `Extrapolate::Error` aggregates out-of-bounds points.
-    fn batch_interpolate_into(
-        &self,
-        data: &InterpData3DBase<D>,
-        points: &[[D::Elem; 3]],
-        out: &mut [D::Elem],
-    ) -> Result<(), InterpolateError> {
-        if out.len() != points.len() {
-            return Err(InterpolateError::OutputLength {
-                expected: points.len(),
-                found: out.len(),
-            });
-        }
-        self.check_batch_domain(points.iter().map(|p| p.as_slice()))?;
-        let transformed_points: Vec<[D::Elem; 3]> = points
-            .iter()
-            .map(|point| core::array::from_fn(|i| self.transforms[i].forward(point[i])))
-            .collect();
-        let values = self.transformed_values_view(data.values.view());
-        let view = InterpData3DView {
-            grid: core::array::from_fn(|i| self.grid_cache[i].view()),
-            values,
-        };
-        self.inner
-            .batch_interpolate_into(&view, &transformed_points, out)
-    }
-
-    /// Checks this layer's own domain per point (not bailing on the first violation),
-    /// then recurses into `inner` with only the *outer-valid* points' forward
-    /// transforms, remapping `inner`'s failure indices back to their true batch
-    /// position: an outer-invalid point can't be forward-transformed meaningfully, but
-    /// its presence must not hide `inner`'s own violations for the *other* points, so
-    /// a `GridTransform` nested inside another one still gets every one of its
-    /// violations aggregated, not just the outer layer's. Exposed here so generic
-    /// callers (e.g. `Extrapolate::Wrap`'s batch dispatch) can pre-scan a batch
-    /// without knowing the concrete strategy type.
-    fn check_batch_domain(&self, points: &[[D::Elem; 3]]) -> Result<(), InterpolateError> {
-        let mut failures = Vec::new();
-        let mut valid_indices = Vec::new();
-        let mut transformed_points: Vec<[D::Elem; 3]> = Vec::new();
-        for (index, point) in points.iter().enumerate() {
-            let point_failures = self.point_domain_failures(index, point.as_slice());
-            if point_failures.is_empty() {
-                valid_indices.push(index);
-                transformed_points.push(core::array::from_fn(|i| {
-                    self.transforms[i].forward(point[i])
-                }));
-            } else {
-                failures.extend(point_failures);
-            }
-        }
-        match self.inner.check_batch_domain(&transformed_points) {
-            Ok(()) => {}
-            Err(InterpolateError::GridTransformDomain(inner_failures)) => {
-                failures.extend(inner_failures.into_iter().map(|f| OutsideDomainAt {
-                    index: valid_indices[f.index],
-                    ..f
-                }));
-            }
-            Err(other) => return Err(other),
-        }
-        failures.sort_by_key(|f| (f.index, f.dim));
-        if failures.is_empty() {
-            Ok(())
-        } else {
-            Err(InterpolateError::GridTransformDomain(failures))
-        }
-    }
-
-    fn allow_extrapolate(&self) -> bool {
-        self.inner.allow_extrapolate()
-    }
-}
-
-impl<D, S> Strategy3D<D> for ValuesTransform<D::Elem, S>
-where
-    D: Data + RawDataClone + Clone,
-    D::Elem: Float + Debug,
-    S: for<'a> Strategy3D<ViewRepr<&'a D::Elem>> + Clone + Debug,
-{
-    /// Checks that every data value is in the configured transform's domain.
-    fn validate(&self, data: &InterpData3DBase<D>) -> Result<(), ValidateError> {
-        let transformed_values = self.transform_values(data.values.view())?;
-        let view = InterpData3DView {
-            grid: core::array::from_fn(|i| data.grid[i].view()),
-            values: transformed_values.view(),
-        };
-        self.inner.validate(&view)
-    }
-
-    /// Transforms `data.values` into `values_cache`, then initializes `inner`
-    /// against a transient view zipping `data.grid` (untouched) with `values_cache`.
-    fn init(&mut self, data: &InterpData3DBase<D>) -> Result<(), ValidateError> {
-        self.values_cache = self.transform_values(data.values.view())?.into_dyn();
-        let view = InterpData3DView {
-            grid: core::array::from_fn(|i| data.grid[i].view()),
-            values: self
-                .values_cache
-                .view()
-                .into_dimensionality::<Ix3>()
-                .expect("values_cache shape matches 3-D values"),
-        };
-        self.inner.init(&view)
-    }
-
-    fn interpolate(
-        &self,
-        data: &InterpData3DBase<D>,
-        point: &[D::Elem; 3],
-    ) -> Result<D::Elem, InterpolateError> {
-        let view = InterpData3DView {
-            grid: core::array::from_fn(|i| data.grid[i].view()),
-            values: self
-                .values_cache
-                .view()
-                .into_dimensionality::<Ix3>()
-                .expect("values_cache shape matches 3-D values"),
-        };
-        let result = self.inner.interpolate(&view, point)?;
-        Ok(self.transform.inverse(result))
-    }
-
-    /// Hands `point` unmodified to `inner.interpolate_wrapped`, so a nested
-    /// `GridTransform` handles the actual raw-space wrap; must not wrap here itself.
-    fn interpolate_wrapped(
-        &self,
-        data: &InterpData3DBase<D>,
-        point: &[D::Elem; 3],
-    ) -> Result<D::Elem, InterpolateError>
-    where
-        D::Elem: Num + Euclid + Copy,
-    {
-        let view = InterpData3DView {
-            grid: core::array::from_fn(|i| data.grid[i].view()),
-            values: self
-                .values_cache
-                .view()
-                .into_dimensionality::<Ix3>()
-                .expect("values_cache shape matches 3-D values"),
-        };
-        let result = self.inner.interpolate_wrapped(&view, point)?;
-        Ok(self.transform.inverse(result))
-    }
-
-    /// Calls `inner.interpolate_fast` rather than `inner.interpolate`, so "fast"
-    /// propagates through nested `GridTransform`/`ValuesTransform` layers instead of
-    /// stopping at the outermost one.
-    fn interpolate_fast(&self, data: &InterpData3DBase<D>, point: &[D::Elem; 3]) -> D::Elem {
-        let view = InterpData3DView {
-            grid: core::array::from_fn(|i| data.grid[i].view()),
-            values: self
-                .values_cache
-                .view()
-                .into_dimensionality::<Ix3>()
-                .expect("values_cache shape matches 3-D values"),
-        };
-        let result = self.inner.interpolate_fast(&view, point);
-        self.transform.inverse(result)
-    }
-
-    /// Delegates the whole batch to `inner` via a transient view over `values_cache`,
-    /// so a nested `GridTransform`'s batch-domain aggregation isn't lost to the
-    /// point-by-point default; inverse-transforms every output afterward.
-    fn batch_interpolate_into(
-        &self,
-        data: &InterpData3DBase<D>,
-        points: &[[D::Elem; 3]],
-        out: &mut [D::Elem],
-    ) -> Result<(), InterpolateError> {
-        if out.len() != points.len() {
-            return Err(InterpolateError::OutputLength {
-                expected: points.len(),
-                found: out.len(),
-            });
-        }
-        let view = InterpData3DView {
-            grid: core::array::from_fn(|i| data.grid[i].view()),
-            values: self
-                .values_cache
-                .view()
-                .into_dimensionality::<Ix3>()
-                .expect("values_cache shape matches 3-D values"),
-        };
-        self.inner.batch_interpolate_into(&view, points, out)?;
-        for o in out.iter_mut() {
-            *o = self.transform.inverse(*o);
-        }
-        Ok(())
-    }
-
-    /// Forwards to `inner`, so a nested `GridTransform`'s domain is still checked
-    /// through a wrapping `ValuesTransform`.
-    fn check_batch_domain(&self, points: &[[D::Elem; 3]]) -> Result<(), InterpolateError> {
-        self.inner.check_batch_domain(points)
-    }
-
-    fn allow_extrapolate(&self) -> bool {
-        self.inner.allow_extrapolate()
-    }
-}
+grid_transform_strategy_impl!(
+    Strategy3D,
+    InterpData3DBase,
+    InterpData3DView,
+    3,
+    "Checks the axis count (must be 1 or 3) and that every raw grid coordinate is \
+     in its axis's configured transform's domain."
+);
+values_transform_strategy_impl!(Strategy3D, InterpData3DBase, InterpData3DView, Ix3, 3);
