+++ src/lib.rs
@@ -58,9 +58,7 @@ pub(crate) use std::fmt::Debug;
 
 pub use ndarray;
 pub(crate) use ndarray::prelude::*;
-pub(crate) use ndarray::{
-    indices_of, Data, IntoDimension, Ix, OwnedRepr, RawDataClone, Slice, ViewRepr,
-};
+pub(crate) use ndarray::{Data, IntoDimension, Ix, OwnedRepr, RawDataClone, Slice, ViewRepr};
 
 pub use num_traits;
 pub(crate) use num_traits::{clamp, Euclid, Float, Num, One, Zero};
