+++ src/strategy/transform.rs
@@ -101,13 +101,6 @@ impl Transform {
     }
 }
 
-/// Placeholder for [`GridTransform::grid_cache`]/[`ValuesTransform::values_cache`]
-/// before `init` populates them. Mirrors [`crate::strategy::cubic`]'s own
-/// `empty_cache`.
-fn empty_cache<T>() -> ArrayD<T> {
-    ArrayD::from_shape_vec(IxDyn(&[0]), Vec::new()).expect("empty shape matches empty vec")
-}
-
 /// Wraps a [`Strategy1D`]/[`Strategy2D`]/[`Strategy3D`]/[`StrategyND`] `inner`
 /// strategy, interpolating in a transformed grid coordinate space (see [`Transform`])
 /// instead of the raw one.
@@ -244,6 +237,7 @@ impl<T: Float, S> GridTransform<T, S> {
         let transform = self.transforms[dim];
         let mut transformed: Vec<T> = Vec::with_capacity(grid.len());
         for (index, &x) in grid.iter().enumerate() {
+            // Domain check.
             if !transform.in_domain(x) {
                 return Err(ValidateError::GridTransformDomain {
                     transform,
@@ -251,7 +245,9 @@ impl<T: Float, S> GridTransform<T, S> {
                     index,
                 });
             }
+            // Forward-transform.
             let fx = transform.forward(x);
+            // Monotonicity check, against the previous transformed coordinate.
             if let Some(&prev) = transformed.last() {
                 let ordered = if transform.is_increasing() {
                     fx > prev
@@ -268,6 +264,7 @@ impl<T: Float, S> GridTransform<T, S> {
             }
             transformed.push(fx);
         }
+        // Reversal, to stay ascending under a decreasing transform.
         let mut transformed = Array1::from_vec(transformed);
         if !transform.is_increasing() {
             transformed.invert_axis(Axis(0));
@@ -342,6 +339,221 @@ impl<T: Float, S> GridTransform<T, S> {
     }
 }
 
+/// Generates `impl $Strategy<D> for GridTransform<D::Elem, S>`, for 1D/2D/3D
+/// (`$N` in `[D::Elem; $N]`); `StrategyND` stays hand-written in `n/strategies.rs`,
+/// since its point type is a runtime-length slice rather than a fixed-size array and
+/// so can't share this shape, mirroring why `StrategyND` itself isn't generated by
+/// [`crate::strategy::traits::fixed_strategy_trait`].
+macro_rules! grid_transform_strategy_impl {
+    ($Strategy:ident, $InterpDataBase:ident, $InterpDataView:ident, $N:literal, $validate_doc:literal) => {
+        impl<D, S> $Strategy<D> for GridTransform<D::Elem, S>
+        where
+            D: Data + RawDataClone + Clone,
+            D::Elem: Float + Debug,
+            S: for<'a> $Strategy<ViewRepr<&'a D::Elem>> + Clone + Debug,
+        {
+            #[doc = $validate_doc]
+            fn validate(&self, data: &$InterpDataBase<D>) -> Result<(), ValidateError> {
+                self.transforms
+                    .validate_len($N, "GridTransform", "transforms")?;
+                let transformed_grid: Vec<Array1<D::Elem>> = data
+                    .grid
+                    .iter()
+                    .enumerate()
+                    .map(|(dim, grid)| self.transform_axis(dim, grid.view()))
+                    .collect::<Result<_, _>>()?;
+                let values = self.transformed_values_view(data.values.view());
+                let view = $InterpDataView {
+                    grid: core::array::from_fn(|i| transformed_grid[i].view()),
+                    values,
+                };
+                self.inner.validate(&view)
+            }
+
+            /// Transforms the grid into `grid_cache`, then initializes `inner` against
+            /// a transient view zipping `grid_cache` with `data.values`.
+            ///
+            /// A raw grid is always strictly increasing, but a decreasing transform
+            /// (e.g. `Reciprocal`) would otherwise leave that axis of `grid_cache`
+            /// decreasing; `Transform::is_increasing` flags that case so the
+            /// transformed axis (and the matching `values` axis) can be reversed back
+            /// to ascending, matching every downstream strategy's ascending-grid
+            /// assumption.
+            fn init(&mut self, data: &$InterpDataBase<D>) -> Result<(), ValidateError> {
+                self.grid_cache = data
+                    .grid
+                    .iter()
+                    .enumerate()
+                    .map(|(dim, grid)| self.transform_axis(dim, grid.view()))
+                    .collect::<Result<_, _>>()?;
+                let values = self.transformed_values_view(data.values.view());
+                let view = $InterpDataView {
+                    grid: core::array::from_fn(|i| self.grid_cache[i].view()),
+                    values,
+                };
+                self.inner.init(&view)
+            }
+
+            fn interpolate(
+                &self,
+                data: &$InterpDataBase<D>,
+                point: &[D::Elem; $N],
+            ) -> Result<D::Elem, InterpolateError> {
+                self.check_point_domain(point)?;
+                let transformed_point: [D::Elem; $N] =
+                    core::array::from_fn(|i| self.transforms[i].forward(point[i]));
+                let values = self.transformed_values_view(data.values.view());
+                let view = $InterpDataView {
+                    grid: core::array::from_fn(|i| self.grid_cache[i].view()),
+                    values,
+                };
+                self.inner.interpolate(&view, &transformed_point)
+            }
+
+            #[doc = concat!(
+                " Forward-transforms into `grid_cache`'s coordinate space, then delegates\n",
+                " the actual wrap to `inner.interpolate_wrapped` rather than wrapping here\n",
+                " itself: wrapping doesn't commute with a nonlinear transform, so it must\n",
+                " happen in the *final* (innermost) transformed space where the periodic\n",
+                " strategy actually lives, mirroring how `ValuesTransform::interpolate_wrapped`\n",
+                " defers to `inner` for the same reason (composing two `GridTransform`s and\n",
+                " wrapping at the outer layer's space, then forward-transforming again, uses\n",
+                " the wrong period). For a non-transform `inner`, `", stringify!($Strategy),
+                "::interpolate_wrapped`'s default wraps directly against `grid_cache`,\n",
+                " reproducing this layer's own wrap exactly.",
+            )]
+            fn interpolate_wrapped(
+                &self,
+                data: &$InterpDataBase<D>,
+                point: &[D::Elem; $N],
+            ) -> Result<D::Elem, InterpolateError>
+            where
+                D::Elem: Num + Euclid + Copy,
+            {
+                self.check_point_domain(point)?;
+                let transformed_point: [D::Elem; $N] =
+                    core::array::from_fn(|i| self.transforms[i].forward(point[i]));
+                let values = self.transformed_values_view(data.values.view());
+                let view = $InterpDataView {
+                    grid: core::array::from_fn(|i| self.grid_cache[i].view()),
+                    values,
+                };
+                self.inner.interpolate_wrapped(&view, &transformed_point)
+            }
+
+            /// Skips the domain check `interpolate` does, and calls
+            /// `inner.interpolate_fast` rather than `inner.interpolate`, so "fast"
+            /// propagates through nested `GridTransform`/`ValuesTransform` layers
+            /// instead of stopping at the outermost one.
+            ///
+            /// An out-of-domain point is the caller's problem here, same as any other
+            /// unchecked `_fast` method: `forward`-ing it produces `NaN` (`Log`/`Sqrt`)
+            /// or `+-inf` (`Reciprocal`, only at `x == 0`). `NaN` poisons the grid
+            /// search's comparisons and panics; `+-inf` compares normally, so it's
+            /// treated like an ordinary out-of-bounds extrapolation query and silently
+            /// produces `NaN` output instead. Expected, not a bug: check
+            /// [`Transform::in_domain`] yourself first if you need a guarantee either
+            /// way.
+            fn interpolate_fast(
+                &self,
+                data: &$InterpDataBase<D>,
+                point: &[D::Elem; $N],
+            ) -> D::Elem {
+                let transformed_point: [D::Elem; $N] =
+                    core::array::from_fn(|i| self.transforms[i].forward(point[i]));
+                let values = self.transformed_values_view(data.values.view());
+                let view = $InterpDataView {
+                    grid: core::array::from_fn(|i| self.grid_cache[i].view()),
+                    values,
+                };
+                self.inner.interpolate_fast(&view, &transformed_point)
+            }
+
+            /// Domain-checks every point in the batch before transforming, aggregating
+            /// every violation across the *whole batch* into one
+            /// [`InterpolateError::GridTransformDomain`] instead of erroring on the
+            /// first one, mirroring how `Extrapolate::Error` aggregates out-of-bounds
+            /// points.
+            fn batch_interpolate_into(
+                &self,
+                data: &$InterpDataBase<D>,
+                points: &[[D::Elem; $N]],
+                out: &mut [D::Elem],
+            ) -> Result<(), InterpolateError> {
+                if out.len() != points.len() {
+                    return Err(InterpolateError::OutputLength {
+                        expected: points.len(),
+                        found: out.len(),
+                    });
+                }
+                self.check_batch_domain(points.iter().map(|p| p.as_slice()))?;
+                let transformed_points: Vec<[D::Elem; $N]> = points
+                    .iter()
+                    .map(|point| core::array::from_fn(|i| self.transforms[i].forward(point[i])))
+                    .collect();
+                let values = self.transformed_values_view(data.values.view());
+                let view = $InterpDataView {
+                    grid: core::array::from_fn(|i| self.grid_cache[i].view()),
+                    values,
+                };
+                self.inner
+                    .batch_interpolate_into(&view, &transformed_points, out)
+            }
+
+            /// Checks this layer's own domain per point (not bailing on the first
+            /// violation), then recurses into `inner` with only the *outer-valid*
+            /// points' forward transforms, remapping `inner`'s failure indices back to
+            /// their true batch position: an outer-invalid point can't be
+            /// forward-transformed meaningfully, but its presence must not hide
+            /// `inner`'s own violations for the *other* points, so a `GridTransform`
+            /// nested inside another one still gets every one of its violations
+            /// aggregated, not just the outer layer's. Exposed here so generic callers
+            /// (e.g. `Extrapolate::Wrap`'s batch dispatch) can pre-scan a batch without
+            /// knowing the concrete strategy type.
+            fn check_batch_domain(
+                &self,
+                points: &[[D::Elem; $N]],
+            ) -> Result<(), InterpolateError> {
+                let mut failures = Vec::new();
+                let mut valid_indices = Vec::new();
+                let mut transformed_points: Vec<[D::Elem; $N]> = Vec::new();
+                for (index, point) in points.iter().enumerate() {
+                    let point_failures = self.point_domain_failures(index, point.as_slice());
+                    if point_failures.is_empty() {
+                        valid_indices.push(index);
+                        transformed_points.push(core::array::from_fn(|i| {
+                            self.transforms[i].forward(point[i])
+                        }));
+                    } else {
+                        failures.extend(point_failures);
+                    }
+                }
+                match self.inner.check_batch_domain(&transformed_points) {
+                    Ok(()) => {}
+                    Err(InterpolateError::GridTransformDomain(inner_failures)) => {
+                        failures.extend(inner_failures.into_iter().map(|f| OutsideDomainAt {
+                            index: valid_indices[f.index],
+                            ..f
+                        }));
+                    }
+                    Err(other) => return Err(other),
+                }
+                failures.sort_by_key(|f| (f.index, f.dim));
+                if failures.is_empty() {
+                    Ok(())
+                } else {
+                    Err(InterpolateError::GridTransformDomain(failures))
+                }
+            }
+
+            fn allow_extrapolate(&self) -> bool {
+                self.inner.allow_extrapolate()
+            }
+        }
+    };
+}
+pub(crate) use grid_transform_strategy_impl;
+
 /// Wraps a [`Strategy1D`]/[`Strategy2D`]/[`Strategy3D`]/[`StrategyND`] `inner`
 /// strategy, interpolating in a transformed *value* space (see [`Transform`])
 /// instead of the raw one: e.g. keeping interpolated output within a range a raw
@@ -407,6 +619,21 @@ impl<T, S> ValuesTransform<T, S> {
     pub fn reciprocal(inner: S) -> Self {
         Self::new(Transform::Reciprocal, inner)
     }
+
+    /// `values_cache` reshaped to `Dim`. Shared by `interpolate`/`interpolate_wrapped`/
+    /// `interpolate_fast`/`batch_interpolate_into` across 1D/2D/3D (ND keeps
+    /// `values_cache` dynamically-shaped throughout, so it doesn't need this).
+    ///
+    /// # Panics
+    /// If the cached shape doesn't match `Dim`. Doesn't happen in practice: `init`
+    /// populates `values_cache` from a `Dim`-shaped `values` array before any of the
+    /// above can run.
+    pub(crate) fn cached_values_view<Dim: Dimension>(&self) -> ArrayView<'_, T, Dim> {
+        self.values_cache
+            .view()
+            .into_dimensionality::<Dim>()
+            .expect("values_cache shape matches values")
+    }
 }
 
 impl<T: Float, S> ValuesTransform<T, S> {
@@ -426,7 +653,7 @@ impl<T: Float, S> ValuesTransform<T, S> {
         values: ArrayView<T, Dim>,
     ) -> Result<Array<T, Dim>, ValidateError> {
         let mut transformed = Vec::with_capacity(values.len());
-        for (pattern, &v) in indices_of(&values).into_iter().zip(values.iter()) {
+        for (pattern, &v) in values.indexed_iter() {
             if !self.transform.in_domain(v) {
                 let index: Dim = pattern.into_dimension();
                 return Err(ValidateError::ValuesTransformDomain {
@@ -440,6 +667,136 @@ impl<T: Float, S> ValuesTransform<T, S> {
             .expect("transformed shape matches values shape"))
     }
 }
+
+/// Generates `impl $Strategy<D> for ValuesTransform<D::Elem, S>`, for 1D/2D/3D
+/// (`$N` in `[D::Elem; $N]`, `$Ix` the matching `ndarray` dimension type);
+/// `StrategyND` stays hand-written in `n/strategies.rs`, for the same reason
+/// [`grid_transform_strategy_impl`] doesn't cover it.
+macro_rules! values_transform_strategy_impl {
+    ($Strategy:ident, $InterpDataBase:ident, $InterpDataView:ident, $Ix:ident, $N:literal) => {
+        impl<D, S> $Strategy<D> for ValuesTransform<D::Elem, S>
+        where
+            D: Data + RawDataClone + Clone,
+            D::Elem: Float + Debug,
+            S: for<'a> $Strategy<ViewRepr<&'a D::Elem>> + Clone + Debug,
+        {
+            /// Checks that every data value is in the configured transform's domain.
+            fn validate(&self, data: &$InterpDataBase<D>) -> Result<(), ValidateError> {
+                let transformed_values = self.transform_values(data.values.view())?;
+                let view = $InterpDataView {
+                    grid: core::array::from_fn(|i| data.grid[i].view()),
+                    values: transformed_values.view(),
+                };
+                self.inner.validate(&view)
+            }
+
+            /// Transforms `data.values` into `values_cache`, then initializes `inner`
+            /// against a transient view zipping `data.grid` (untouched) with
+            /// `values_cache`.
+            fn init(&mut self, data: &$InterpDataBase<D>) -> Result<(), ValidateError> {
+                self.values_cache = self.transform_values(data.values.view())?.into_dyn();
+                // Not `cached_values_view`: that borrows all of `self`, conflicting
+                // with `self.inner.init`'s `&mut self.inner` below. Direct field
+                // access keeps the two borrows disjoint.
+                let view = $InterpDataView {
+                    grid: core::array::from_fn(|i| data.grid[i].view()),
+                    values: self
+                        .values_cache
+                        .view()
+                        .into_dimensionality::<$Ix>()
+                        .expect("values_cache shape matches values"),
+                };
+                self.inner.init(&view)
+            }
+
+            fn interpolate(
+                &self,
+                data: &$InterpDataBase<D>,
+                point: &[D::Elem; $N],
+            ) -> Result<D::Elem, InterpolateError> {
+                let view = $InterpDataView {
+                    grid: core::array::from_fn(|i| data.grid[i].view()),
+                    values: self.cached_values_view(),
+                };
+                let result = self.inner.interpolate(&view, point)?;
+                Ok(self.transform.inverse(result))
+            }
+
+            /// Hands `point` unmodified to `inner.interpolate_wrapped`, so a nested
+            /// `GridTransform` handles the actual raw-space wrap; must not wrap here
+            /// itself.
+            fn interpolate_wrapped(
+                &self,
+                data: &$InterpDataBase<D>,
+                point: &[D::Elem; $N],
+            ) -> Result<D::Elem, InterpolateError>
+            where
+                D::Elem: Num + Euclid + Copy,
+            {
+                let view = $InterpDataView {
+                    grid: core::array::from_fn(|i| data.grid[i].view()),
+                    values: self.cached_values_view(),
+                };
+                let result = self.inner.interpolate_wrapped(&view, point)?;
+                Ok(self.transform.inverse(result))
+            }
+
+            /// Calls `inner.interpolate_fast` rather than `inner.interpolate`, so
+            /// "fast" propagates through nested `GridTransform`/`ValuesTransform`
+            /// layers instead of stopping at the outermost one.
+            fn interpolate_fast(
+                &self,
+                data: &$InterpDataBase<D>,
+                point: &[D::Elem; $N],
+            ) -> D::Elem {
+                let view = $InterpDataView {
+                    grid: core::array::from_fn(|i| data.grid[i].view()),
+                    values: self.cached_values_view(),
+                };
+                let result = self.inner.interpolate_fast(&view, point);
+                self.transform.inverse(result)
+            }
+
+            /// Delegates the whole batch to `inner` via a transient view over
+            /// `values_cache`, so a nested `GridTransform`'s batch-domain aggregation
+            /// isn't lost to the point-by-point default; inverse-transforms every
+            /// output afterward.
+            fn batch_interpolate_into(
+                &self,
+                data: &$InterpDataBase<D>,
+                points: &[[D::Elem; $N]],
+                out: &mut [D::Elem],
+            ) -> Result<(), InterpolateError> {
+                if out.len() != points.len() {
+                    return Err(InterpolateError::OutputLength {
+                        expected: points.len(),
+                        found: out.len(),
+                    });
+                }
+                let view = $InterpDataView {
+                    grid: core::array::from_fn(|i| data.grid[i].view()),
+                    values: self.cached_values_view(),
+                };
+                self.inner.batch_interpolate_into(&view, points, out)?;
+                for o in out.iter_mut() {
+                    *o = self.transform.inverse(*o);
+                }
+                Ok(())
+            }
+
+            /// Forwards to `inner`, so a nested `GridTransform`'s domain is still
+            /// checked through a wrapping `ValuesTransform`.
+            fn check_batch_domain(&self, points: &[[D::Elem; $N]]) -> Result<(), InterpolateError> {
+                self.inner.check_batch_domain(points)
+            }
+
+            fn allow_extrapolate(&self) -> bool {
+                self.inner.allow_extrapolate()
+            }
+        }
+    };
+}
+pub(crate) use values_transform_strategy_impl;
 
 #[cfg(test)]
 mod tests {
